changed to new approach for parseJSON , ...
moves from left to right thru the json string building either single
array or object as it progresses. if nesting occurs (not yet implemented)
possibly a reference to the parent object get's passed in the recursive call.
Or maybe not,

obj1 = {"a": {"b": true} }

system detects the second '{' and knows to set for recursion:

detected '{ ... {' so prep the key and object for recursion
obj["a"] = parseJSON(s, p, o);

the third parameter for the object might not be needed at all...